+++ mex/editor/aux_search/main.py
@@ -22,13 +22,17 @@ def expand_properties_button(result: AuxResult, index: int) -> rx.Component:
     )
 
 
-def import_button(index: int) -> rx.Component:
+def import_button(result: AuxResult, index: int) -> rx.Component:
     """Render a button to import the aux search result to the MEx backend."""
-    return rx.button(
-        "Import",
-        on_click=lambda: AuxState.import_result(index),
-        align="end",
-        disabled=True,
+    return rx.cond(
+        result.show_import_button,
+        rx.button(
+            "Import",
+            align="end",
+        ),
+        rx.button(
+            "Import", on_click=AuxState.import_result(index), align="end", disabled=True
+        ),
     )
 
 
@@ -87,7 +91,7 @@ def result_title_and_buttons(result: AuxResult, index: int) -> rx.Component:
             },
         ),
         expand_properties_button(result, index),
-        import_button(index),
+        import_button(result, index),
         style={"width": "100%"},
     )