+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- button to import and ingest wikidata organizations
 - add color mode toggle to login screen
 - added new/remove buttons for additive rules for string fields
 - add event handlers for adding/removing additive value input cards